+++ beacon-chain/db/kv/lightclient.go
@@ -7,6 +7,8 @@ import (
 
 	"github.com/golang/snappy"
 	"github.com/pkg/errors"
+	fieldparams "github.com/prysmaticlabs/prysm/v5/config/fieldparams"
+	"github.com/prysmaticlabs/prysm/v5/config/params"
 	"github.com/prysmaticlabs/prysm/v5/consensus-types/interfaces"
 	light_client "github.com/prysmaticlabs/prysm/v5/consensus-types/light-client"
 	"github.com/prysmaticlabs/prysm/v5/encoding/bytesutil"
@@ -35,9 +37,35 @@ func (s *Store) SaveLightClientBootstrap(ctx context.Context, blockRoot []byte, 
 	_, span := trace.StartSpan(ctx, "BeaconDB.SaveLightClientBootstrap")
 	defer span.End()
 
+	bootstrapCopy, err := light_client.NewWrappedBootstrap(proto.Clone(bootstrap.Proto()))
+	if err != nil {
+		return errors.Wrap(err, "could not clone light client bootstrap")
+	}
+	syncCommitteeHash, err := bootstrapCopy.CurrentSyncCommittee().HashTreeRoot()
+	if err != nil {
+		return errors.Wrap(err, "could not hash current sync committee")
+	}
+
 	return s.db.Update(func(tx *bolt.Tx) error {
+		syncCommitteeBucket := tx.Bucket(lightClientSyncCommitteeBucket)
+		syncCommitteeAlreadyExists := syncCommitteeBucket.Get(syncCommitteeHash[:]) != nil
+		if !syncCommitteeAlreadyExists {
+			enc, err := bootstrapCopy.CurrentSyncCommittee().MarshalSSZ()
+			if err != nil {
+				return errors.Wrap(err, "could not marshal current sync committee")
+			}
+			if err := syncCommitteeBucket.Put(syncCommitteeHash[:], enc); err != nil {
+				return errors.Wrap(err, "could not save current sync committee")
+			}
+		}
+
+		err = bootstrapCopy.SetCurrentSyncCommittee(createEmptySyncCommittee())
+		if err != nil {
+			return errors.Wrap(err, "could not set current sync committee to zero while saving")
+		}
+
 		bkt := tx.Bucket(lightClientBootstrapBucket)
-		enc, err := encodeLightClientBootstrap(bootstrap)
+		enc, err := encodeLightClientBootstrap(bootstrapCopy, syncCommitteeHash)
 		if err != nil {
 			return err
 		}
@@ -50,20 +78,49 @@ func (s *Store) LightClientBootstrap(ctx context.Context, blockRoot []byte) (int
 	defer span.End()
 
 	var bootstrap interfaces.LightClientBootstrap
+	var syncCommitteeHash []byte
 	err := s.db.View(func(tx *bolt.Tx) error {
 		bkt := tx.Bucket(lightClientBootstrapBucket)
+		syncCommitteeBucket := tx.Bucket(lightClientSyncCommitteeBucket)
 		enc := bkt.Get(blockRoot)
 		if enc == nil {
 			return nil
 		}
 		var err error
-		bootstrap, err = decodeLightClientBootstrap(enc)
+		bootstrap, syncCommitteeHash, err = decodeLightClientBootstrap(enc)
+		if err != nil {
+			return errors.Wrap(err, "could not decode light client bootstrap")
+		}
+		var syncCommitteeBytes = syncCommitteeBucket.Get(syncCommitteeHash)
+		if syncCommitteeBytes == nil {
+			return errors.New("sync committee not found")
+		}
+		syncCommittee := &ethpb.SyncCommittee{}
+		if err := syncCommittee.UnmarshalSSZ(syncCommitteeBytes); err != nil {
+			return errors.Wrap(err, "could not unmarshal sync committee")
+		}
+		err = bootstrap.SetCurrentSyncCommittee(syncCommittee)
+		if err != nil {
+			return errors.Wrap(err, "could not set current sync committee while retrieving")
+		}
 		return err
 	})
 	return bootstrap, err
 }
 
-func encodeLightClientBootstrap(bootstrap interfaces.LightClientBootstrap) ([]byte, error) {
+func createEmptySyncCommittee() *ethpb.SyncCommittee {
+	syncCom := make([][]byte, params.BeaconConfig().SyncCommitteeSize)
+	for i := 0; uint64(i) < params.BeaconConfig().SyncCommitteeSize; i++ {
+		syncCom[i] = make([]byte, fieldparams.BLSPubkeyLength)
+	}
+
+	return &ethpb.SyncCommittee{
+		Pubkeys:         syncCom,
+		AggregatePubkey: make([]byte, fieldparams.BLSPubkeyLength),
+	}
+}
+
+func encodeLightClientBootstrap(bootstrap interfaces.LightClientBootstrap, syncCommitteeHash [32]byte) ([]byte, error) {
 	key, err := keyForLightClientUpdate(bootstrap.Version())
 	if err != nil {
 		return nil, err
@@ -72,48 +129,56 @@ func encodeLightClientBootstrap(bootstrap interfaces.LightClientBootstrap) ([]by
 	if err != nil {
 		return nil, errors.Wrap(err, "could not marshal light client bootstrap")
 	}
-	fullEnc := make([]byte, len(key)+len(enc))
+	fullEnc := make([]byte, len(key)+32+len(enc))
 	copy(fullEnc, key)
-	copy(fullEnc[len(key):], enc)
-	return snappy.Encode(nil, fullEnc), nil
+	copy(fullEnc[len(key):len(key)+32], syncCommitteeHash[:])
+	copy(fullEnc[len(key)+32:], enc)
+	compressedEnc := snappy.Encode(nil, fullEnc)
+	return compressedEnc, nil
 }
 
-func decodeLightClientBootstrap(enc []byte) (interfaces.LightClientBootstrap, error) {
+func decodeLightClientBootstrap(enc []byte) (interfaces.LightClientBootstrap, []byte, error) {
 	var err error
 	enc, err = snappy.Decode(nil, enc)
 	if err != nil {
-		return nil, errors.Wrap(err, "could not snappy decode light client bootstrap")
+		return nil, nil, errors.Wrap(err, "could not snappy decode light client bootstrap")
 	}
 	var m proto.Message
+	var syncCommitteeHash []byte
 	switch {
 	case hasAltairKey(enc):
 		bootstrap := &ethpb.LightClientBootstrapAltair{}
-		if err := bootstrap.UnmarshalSSZ(enc[len(altairKey):]); err != nil {
-			return nil, errors.Wrap(err, "could not unmarshal Altair light client bootstrap")
+		if err := bootstrap.UnmarshalSSZ(enc[len(altairKey)+32:]); err != nil {
+			return nil, nil, errors.Wrap(err, "could not unmarshal Altair light client bootstrap")
 		}
 		m = bootstrap
+		syncCommitteeHash = enc[len(altairKey) : len(altairKey)+32]
 	case hasCapellaKey(enc):
 		bootstrap := &ethpb.LightClientBootstrapCapella{}
-		if err := bootstrap.UnmarshalSSZ(enc[len(capellaKey):]); err != nil {
-			return nil, errors.Wrap(err, "could not unmarshal Capella light client bootstrap")
+		if err := bootstrap.UnmarshalSSZ(enc[len(capellaKey)+32:]); err != nil {
+			return nil, nil, errors.Wrap(err, "could not unmarshal Capella light client bootstrap")
 		}
 		m = bootstrap
+		syncCommitteeHash = enc[len(capellaKey) : len(capellaKey)+32]
 	case hasDenebKey(enc):
 		bootstrap := &ethpb.LightClientBootstrapDeneb{}
-		if err := bootstrap.UnmarshalSSZ(enc[len(denebKey):]); err != nil {
-			return nil, errors.Wrap(err, "could not unmarshal Deneb light client bootstrap")
+		if err := bootstrap.UnmarshalSSZ(enc[len(denebKey)+32:]); err != nil {
+			return nil, nil, errors.Wrap(err, "could not unmarshal Deneb light client bootstrap")
 		}
 		m = bootstrap
+		syncCommitteeHash = enc[len(denebKey) : len(denebKey)+32]
 	case hasElectraKey(enc):
 		bootstrap := &ethpb.LightClientBootstrapElectra{}
-		if err := bootstrap.UnmarshalSSZ(enc[len(electraKey):]); err != nil {
-			return nil, errors.Wrap(err, "could not unmarshal Electra light client bootstrap")
+		if err := bootstrap.UnmarshalSSZ(enc[len(electraKey)+32:]); err != nil {
+			return nil, nil, errors.Wrap(err, "could not unmarshal Electra light client bootstrap")
 		}
 		m = bootstrap
+		syncCommitteeHash = enc[len(electraKey) : len(electraKey)+32]
 	default:
-		return nil, errors.New("decoding of saved light client bootstrap is unsupported")
+		return nil, nil, errors.New("decoding of saved light client bootstrap is unsupported")
 	}
-	return light_client.NewWrappedBootstrap(m)
+	bootstrap, err := light_client.NewWrappedBootstrap(m)
+	return bootstrap, syncCommitteeHash, err
 }
 
 func (s *Store) LightClientUpdates(ctx context.Context, startPeriod, endPeriod uint64) (map[uint64]interfaces.LightClientUpdate, error) {