+++ beacon-chain/db/kv/schema.go
@@ -18,8 +18,9 @@ var (
 	registrationBucket    = []byte("registration")
 
 	// Light Client Updates Bucket
-	lightClientUpdatesBucket   = []byte("light-client-updates")
-	lightClientBootstrapBucket = []byte("light-client-bootstrap")
+	lightClientUpdatesBucket       = []byte("light-client-updates")
+	lightClientBootstrapBucket     = []byte("light-client-bootstrap")
+	lightClientSyncCommitteeBucket = []byte("light-client-sync-committee")
 
 	// Deprecated: This bucket was migrated in PR 6461. Do not use, except for migrations.
 	slotsHasObjectBucket = []byte("slots-has-objects")